PostResource working
-Added new postgres container
-Changed db initialization to only when there are pending migrations

+++ DAPM/DAPM.ResourceRegistryMS.Api/Data/ResourceRegistryDbContext.cs
@@ -24,39 +24,33 @@ public class ResourceRegistryDbContext : DbContext
 
     public void InitializeDatabase()
     {
-        Database.EnsureDeleted();
-        Database.Migrate();
+        if (Database.GetPendingMigrations().Any())
+        {
+            Database.EnsureDeleted();
+            Database.Migrate();
 
-        ResourceType csv = new ResourceType { Name = "EventLog", FileExtension = ".csv" };
-        ResourceTypes.Add(csv);
+            ResourceType csv = new ResourceType { Name = "EventLog", FileExtension = ".csv" };
+            ResourceTypes.Add(csv);
 
-        Peer dtuPeer = new Peer { Name = "DTU", ApiUrl = "http://dtu.dk" };
-        Peer kuPeer = new Peer { Name = "KU", ApiUrl = "http://ku.dk" };
-        Peer nvPeer = new Peer { Name = "Novo Nordisk", ApiUrl = "http://novonordisk.dk" };
+            Peer dtuPeer = new Peer { Name = "DTU", ApiUrl = "http://dtu.dk" };
+            Peer kuPeer = new Peer { Name = "KU", ApiUrl = "http://ku.dk" };
+            Peer nvPeer = new Peer { Name = "Novo Nordisk", ApiUrl = "http://novonordisk.dk" };
 
-        Peers.Add(dtuPeer);
-        Peers.Add(kuPeer);
-        Peers.Add(nvPeer);
+            Peers.Add(dtuPeer);
+            Peers.Add(kuPeer);
+            Peers.Add(nvPeer);
 
 
-        Repository dtuRepo = new Repository { Name = "DTU Repository", Peer = dtuPeer };
-        Repository kuRepo = new Repository { Name = "KU Repository", Peer = kuPeer };
-        Repository nvRepo = new Repository { Name = "NovoNordisk Repository", Peer = nvPeer };
+            Repository dtuRepo = new Repository { Name = "DTU Repository", Peer = dtuPeer };
+            Repository kuRepo = new Repository { Name = "KU Repository", Peer = kuPeer };
+            Repository nvRepo = new Repository { Name = "NovoNordisk Repository", Peer = nvPeer };
 
-        Repositories.Add(dtuRepo);
-        Repositories.Add(kuRepo);
-        Repositories.Add(nvRepo);
+            Repositories.Add(dtuRepo);
+            Repositories.Add(kuRepo);
+            Repositories.Add(nvRepo);
 
-        Resource dtuResource = new Resource { Name = "EventLogDtu", Repository = dtuRepo, ResourceType = csv };
-        Resource kuResource = new Resource { Name = "EventLogKu", Repository = kuRepo, ResourceType = csv };
-        Resource nvResource = new Resource { Name = "EventLogNn", Repository = nvRepo, ResourceType = csv };
 
-        Resources.Add(dtuResource);
-        Resources.Add(kuResource);
-        Resources.Add(nvResource);
-
-        SaveChanges();
+            SaveChanges();
+        }
     }
-
-
 }
